chore: use express to ignite the bot
