+++ package.json
@@ -31,7 +31,7 @@
     "country-region-data": "^3.0.0",
     "country-state-city": "^3.2.1",
     "dotenv": "^16.4.5",
-    "express": "^4.18.3",
+    "express": "^4.19.2",
     "helmet": "^7.1.0",
     "jest": "^29.7.0",
     "mongoose": "^8.2.1",
