+++ src/loaders/bot.ts
@@ -1,35 +1,38 @@
-import { Telegraf, Context, session, MiddlewareFn } from 'telegraf';
+import { Telegraf, Context, session, Scenes } from 'telegraf';
 import config from '../config/config';
 import dbConnecion from './db-connecion';
+
+import RegistrationScene from '../modules/registration/registration.scene';
+import { checkAndRedirectToScene } from '../middleware/check-command';
+import { checkUserInChannelandPromtJoin } from '../middleware/auth';
+import MainmenuScene from '../modules/mainmenu/mainmenu.scene';
+import Service1Scene from '../modules/service1/service1.scene';
+import QuestionPostScene from '../modules/question-post/question-post.scene';
+import ProfileScene from '../modules/profile/profile.scene';
+import { setCommands } from '../utils/helper/commands';
 
 let bot: Telegraf<Context> | null = null;
 
 export default () => {
   if (bot != null) return bot;
   bot = new Telegraf(config.bot_token as string);
-  dbConnecion;
-  try {
-    bot
-      .launch({
-        webhook: {
-          domain: config.domain,
-          port: 8080,
-        },
-      })
-      .then(() => {
-        // 6668727233
-        // if (bot) console.log(bot);
-        // bot?.context
-        // return bot.sendMessage('6668727233', 'data', {
-        //   parse_mode: 'HTML',
-        //   reply_markup: { remove_keyboard: true },
-        // });
+  bot.telegram.setWebhook(`${config.domain}/secret-path`);
+  const stage = new Scenes.Stage([RegistrationScene, MainmenuScene, Service1Scene, ProfileScene, ...QuestionPostScene]);
+  bot.use(session());
+  bot.use(stage.middleware());
+  bot.use(checkUserInChannelandPromtJoin());
+  bot.use(checkAndRedirectToScene());
 
-        console.log('bot is running');
-      });
-    return bot;
-  } catch (error) {
-    console.log(error);
-    return null;
-  }
+  // Display help with commands and descriptions
+  const commands = [
+    { name: 'start', description: 'Start the bot' },
+    { name: 'help', description: 'Display help' },
+    { name: 'menu', description: 'Display main menu' },
+    { name: 'register', description: 'Register to the bot' },
+    { name: 'profile', description: 'View your profile' },
+    // { name: 'question', description: 'Post a question' },
+  ];
+  setCommands(commands);
+  dbConnecion;
+  return bot;
 };
